Remove /etc/hosts hostname loopback entry
In addition to the 'localhost' entry, some distributions add a loopback
entry for the local hostname in the /etc/hosts file. This is done as a
failsafe in case DNS is not able to resolve the local hostname.

Unfortunately these /etc/hosts entries can interfer with our cell setup,
since we don't wan to use loopback addresses in the VLDB, so just remove
these problematic /etc/hosts entries in the openafs_server role.