Add openafs_add_keys module to set the service keys
Currently we try to set the service keys on each server by trying to run
kadmin on each server, which is confusing and error prone. Also, we use
the new akeyconvert to import the keys, but akeyconvert is not available
on older versions or is packaged on older platforms.

Instead, assume a service key has been created and exported to a keytab
file before the server role runs, and provide a new module to
idempotently import the keys with the 'asetkey' utility.

The openafs_add_keys module detects if the asetkey program supports
extended keys by running it and checking the usage output. The module
will also decode the keytab to find the kvno and enctypes needed for the
asetkey arguments.  (The keytab is decoded directly to avoid a klist
dependency on the server.)

If the keytab is named 'rxkad.keytab' and contains no DES keys and
akeytab does not support extended keys, then the module will just assume
the server has rxkad-k5 support.