Add afs-scenario to create templated molecule scenarios
Add the afs-scenario tool to create templated molecule scenarios so
developers can customize the molecule driver and related settings
depending on the hypervisor and images used for testing.

The afs-scenario tool has three subcommands:

init - Create the initial files for a new molecule scenario using the
built-in cookiecutter templates. The developer should edit and commit
the files as needed after running init. Instead of creating a molecule
file, init will create a molecule.yml.j2 template to support different
drivers.

configure - Create a driver specification file (driver.yml) which
indicates the driver name and options to be used on the test system.
This will create a yaml file which can be stored on the developer's
system.

render - Create the molecule.yml files from the template and the
driver.yml file. Regular molecule commands can be run after the render
command has been completed.